refactor(tests): consolidate contract test helpers
